fix: 댓글 삭제 시 유효성 검증 조건 변경

+++ src/main/java/com/example/MutsaMarket/service/CommentService.java
@@ -74,7 +74,7 @@ public class CommentService {
         if(!itemId.equals(comment.getItemId()))
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
 
-        if(comment.getPassword().equals(dto.getPassword()))
+        if(comment.getWriter().equals(dto.getWriter()) && comment.getPassword().equals(dto.getPassword()))
             commentRepository.deleteById(commentId);
         else
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
